Use coerced type in inlist expr planning
